Node mergins with arbitrary child numbers
Merges of tree lengths can now be done for arbitrary numbers of child
nodes. The node struct still only supports three edges, but the length
merge will work with 1, 2, or 3 child nodes. If there are no children
then this will set the length_determined to false and the data within
should then be ignored in upstream merges.